refactor: extract population helpers for actions panel
