+++ packages/web/src/components/actions/ActionsPanel.tsx
@@ -22,6 +22,13 @@ import { getRequirementIcons } from '../../utils/getRequirementIcons';
 import { useAnimate } from '../../utils/useAutoAnimate';
 import GenericActions from './GenericActions';
 import { formatMissingResources } from './utils';
+import {
+	isHirablePopulation,
+	collectPopulationRolesFromEffects,
+	getPopulationIconFromRole,
+	type EffectConfig,
+	type PopulationRegistryLike,
+} from './populationHelpers';
 
 export interface Action {
 	id: string;
@@ -47,30 +54,9 @@ interface Building {
 	focus?: Focus;
 }
 
-interface PopulationDefinition {
-	id: string;
-	name: string;
-	icon?: string;
-	upkeep?: Record<string, number>;
-	onAssigned?: unknown[];
-	onUnassigned?: unknown[];
-	onGrowthPhase?: unknown[];
-	onUpkeepPhase?: unknown[];
-	onPayUpkeepStep?: unknown[];
-	onGainIncomeStep?: unknown[];
-	onGainAPStep?: unknown[];
-}
-
 export type DisplayPlayer = ReturnType<
 	typeof useGameEngine
 >['ctx']['activePlayer'];
-
-type EffectConfig = {
-	type?: string;
-	method?: string;
-	params?: Record<string, unknown>;
-	effects?: EffectConfig[];
-};
 
 type RequirementConfig = {
 	type?: string;
@@ -82,82 +68,6 @@ type EvaluatorConfig = {
 	type?: string;
 	params?: Record<string, unknown>;
 };
-
-type PopulationRegistryLike = {
-	get(id: string): PopulationDefinition;
-	entries(): [string, PopulationDefinition][];
-};
-
-function isHirablePopulation(
-	population: PopulationDefinition | undefined,
-): boolean {
-	if (!population) {
-		return false;
-	}
-	if (population.upkeep && Object.keys(population.upkeep).length > 0) {
-		return true;
-	}
-	const effectLists: (keyof PopulationDefinition)[] = [
-		'onAssigned',
-		'onUnassigned',
-		'onGrowthPhase',
-		'onUpkeepPhase',
-		'onPayUpkeepStep',
-		'onGainIncomeStep',
-		'onGainAPStep',
-	];
-	return effectLists.some((key) => {
-		const effects = population[key];
-		return Array.isArray(effects) && effects.length > 0;
-	});
-}
-
-function collectPopulationRolesFromEffects(
-	effects: EffectConfig[] | undefined,
-	explicitRoles: Set<string>,
-): boolean {
-	let usesPlaceholder = false;
-	for (const effect of effects ?? []) {
-		if (effect.type === 'population' && effect.method === 'add') {
-			const role = effect.params?.['role'];
-			if (typeof role === 'string') {
-				if (role.startsWith('$')) {
-					usesPlaceholder = true;
-				} else {
-					explicitRoles.add(role);
-				}
-			}
-		}
-		if (effect.effects?.length) {
-			if (collectPopulationRolesFromEffects(effect.effects, explicitRoles)) {
-				usesPlaceholder = true;
-			}
-		}
-	}
-	return usesPlaceholder;
-}
-
-function getPopulationIconFromRole(
-	role: string,
-	populations: PopulationRegistryLike,
-): string {
-	if (!role) {
-		return '';
-	}
-	const infoIcon = POPULATION_ROLES[role as PopulationRoleId]?.icon;
-	if (infoIcon) {
-		return infoIcon;
-	}
-	try {
-		const population = populations.get(role);
-		if (typeof population?.icon === 'string') {
-			return population.icon;
-		}
-	} catch {
-		// Ignore missing population entries when deriving icons.
-	}
-	return '';
-}
 
 function getIconsFromEvaluator(
 	evaluator: EvaluatorConfig | undefined,
